requirements.txt
Flask>=2.0.0
Flask-Cors>=4.0.0
Flask-Login>=0.6.0
werkzeug>=2.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
nltk>=3.6.0
python-dateutil>=2.8.2
spacy>=3.0.0
numpy>=1.21.0
python-dotenv>=1.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
psycopg2-binary>=2.9.0
transformers>=4.21.0
torch>=1.12.0
sentence-transformers>=2.2.0
bcrypt>=4.0.0
faiss-cpu>=1.7.0
openpyxl>=3.0.0
xlrd>=2.0.0